Fix build process
- Move cpp files into src
- Fix by removing unity in main
