+++ platformio.ini
@@ -16,7 +16,7 @@ lib_deps =
     https://github.com/G1-H25/Communication-Library
     ThrowTheSwitch/unity @ ^2.6.0
 board_upload.flash_size = 4MB
-build_flags = -Iinclude -Ilib -DARDUINO_USB_CDC_ON_BOOT=1 -DBOARD_HAS_PSRAM=4MB
+build_flags = -Iinclude -DBOARD_HAS_PSRAM=1
 
 [env:esp32_test]
 platform = espressif32
@@ -25,12 +25,12 @@ framework = espidf
 test_framework = unity
 build_src_filter = +<*/**.cpp> -<main.cpp>
 test_build_src = yes
-test_filter = board/*
+test_filter = board/**
 lib_deps =
     https://github.com/G1-H25/Communication-Library
     ThrowTheSwitch/unity @ ^2.6.0
 board_upload.flash_size = 4MB
-build_flags = -Iinclude -Ilib -DARDUINO_USB_CDC_ON_BOOT=1 -DBOARD_HAS_PSRAM=4MB
+build_flags = -Iinclude -DBOARD_HAS_PSRAM=1
 ; extra_scripts = post:extra_script.py
 
 
@@ -38,23 +38,23 @@ build_flags = -Iinclude -Ilib -DARDUINO_USB_CDC_ON_BOOT=1 -DBOARD_HAS_PSRAM=4MB
 platform = native
 test_framework = unity
 test_build_src = yes
-build_src_filter = -<main.cpp> +<*/**.cpp>
+build_src_filter = -<main.cpp> +<*/**.cpp> -<src/storage/flash_buffer.cpp>
 test_filter = native/unit/*
 lib_deps =
     https://github.com/G1-H25/Communication-Library
     ThrowTheSwitch/unity @ ^2.6.0
-build_flags = -Iinclude -Ilib
+build_flags = -Iinclude
 
 [env:native_smoke]
 platform = native
 test_framework = unity
-test_build_src = no
-build_src_filter = -<*>
+test_build_src = yes
+build_src_filter = -<main.cpp> +<*/**.cpp> -<src/storage/flash_buffer.cpp>
 test_filter = native/smoke/*
 lib_deps =
     https://github.com/G1-H25/Communication-Library
     ThrowTheSwitch/unity @ ^2.6.0
-build_flags = -Iinclude -Ilib
+build_flags = -Iinclude
 
 [env:native_all]
 platform = native
@@ -65,4 +65,4 @@ test_filter = native/*
 lib_deps =
     https://github.com/G1-H25/Communication-Library
     ThrowTheSwitch/unity @ ^2.6.0
-build_flags = -Iinclude -Ilib
+build_flags = -Iinclude
