+++ src/CMakeLists.txt
@@ -1,7 +1,5 @@
-# This file was automatically generated for projects
-# without default 'CMakeLists.txt' file.
+# Check all files in /src and /include and add them to the project
 
-FILE(GLOB_RECURSE app_sources ${CMAKE_SOURCE_DIR}/src/*.*)
-FILE(GLOB_RECURSE app_sources ${CMAKE_SOURCE_DIR}/include/**/*.*)
+FILE(GLOB_RECURSE app_sources ${CMAKE_SOURCE_DIR}/src/*.cpp)
 
-idf_component_register(SRCS ${app_sources})
+idf_component_register(SRCS ${app_sources} INCLUDE_DIRS ${CMAKE_SOURCE_DIR}/include)
